streamlit/requirements.txt
streamlit==1.29.0
psycopg2-binary==2.9.9
pandas==2.1.4
python-dotenv==1.0.0
